Tree Cover Processing Chain

This function completley chains the pre-processing functionality for GFW data from a baseline tree cover percentage raster nad raw loss year and co2 emission information and returns a RasterStack object clipped to the respective aoi including the raw data files as well as the threshold applied versions in addition to yearly tree coverage layers.

prepTC_large(
  inputForestMap,
  inputLossMap,
  inputCO2Map,
  studysite,
  polyName,
  proj,
  bufferRadius,
  thresholdCover,
  thresholdClump,
  ncores,
  outdir,
  tmpdir,
  years,
  overwrite = FALSE
)

Arguments

inputForestMap
A Raster object representing forest cover.

inputLossMap
A RasterLayer object with integer values indicatin the year forest loss ouccoured.

inputCO2Map
A RasterLayer object with the CO2 emission value of total tree cover loss for a given pixel.

studysite
An sf object with the polygons for which to apply the pre- processing chain.

polyName
A charachter vector inicating a column in the sf object which uniquley identifies the polygons.

proj
A charachter vector representing an EPSG code reffering to a coordinated reference system the data is projected to.

bufferRadius
An numeric of length one indicating the size of the buffer in map units.

thresholdCover
A numeric vector of length one in the range of

thresholdClump
A numeric of length one indicating the number of

ncores
A numeric of length one indicating the number of cores used for parallel processing.

outdir
A charachter vector of length one pointing to a directory where the raster output is written to.

tmpdir
A charachter vector of length one pointing to a directory where temporary files can be writte to. This has to be a different directory then specified in outdir, because intermediate files are removed based on the name tag resulting from the values in in the polyName column. All temporary files are deleted.

years
A vector of type numeric indicating the years which are represented by pixels. For GFW data and the default these are 2001 to 2018.

overwrite
A logical indicating if existing files in the outdir directory should be overwritten when the names are equal. Defaults to FALSE.

Value

A charachter vector pointing to the resulting raster files in outdir each matching a polygon object in the studysite object.