Available indicators

Provides the set of indicators available for the supplied input data

Usage

available_indicators(
  stand_static_input = NULL,
  stand_dynamic_input = NULL,
  plant_static_input = NULL,
  plant_dynamic_input = NULL
)

Arguments

stand_static_input

A data frame (or sf object) containing static stand variables. Minimum required column are id_stand (character).

stand_dynamic_input

Optional data frame containing dynamic stand variables. Minimum required columns are id_stand (character) and date (Date).

plant_static_input

Optional data frame containing static plant variables. Minimum required columns are id_stand (character) and plant_entity (character).

plant_dynamic_input

Optional data frame containing dynamic plant variables. Minimum required columns are id_stand (character), plant_entity (character) and date (Date).

Value

A character string with the set of indicators that can be estimated given the supplied input data.

See also

Examples

available_indicators(plant_dynamic_input = example_plant_dynamic_input)
#> [1] "timber_harvest"    "carbon_stock"      "density_dead_wood"